requirements.txt
python_resize_image==1.1.19
requests==2.22.0
instabot==0.70.0
Pillow==6.2.1
python-dotenv==0.10.3
